refactor: 플레이어와 딜러의 공통된 기능 추상 클래스로 변경
